Limitación del tamaño de imagen

+++ app/app.py
@@ -8,8 +8,8 @@ from uuid import uuid4
 import base64
 from model import lightUp
 
-HOST='localhost'
-PORT_NUMBER='8080'
+#HOST='localhost'
+#PORT_NUMBER='8080'
 
 UPLOAD_FOLDER = 'upload'
 
@@ -39,15 +39,20 @@ def light_up():
         image=request.files['file']
         extension=image.filename.split('.')[-1]
         if extension not in ['jpeg','jpg','png']:
-        	flash('Unsupported format :(  Supported formats are .jpg, .jpeg and .png')
+        	flash('Unsupported format :( - Supported formats are .jpg, .jpeg and .png')
         	return render_template('lightup-input.html')
         identifier=uuid4()
         path = os.path.join(app.config['UPLOAD_FOLDER'],str(identifier)+'-'+image.filename)
         image.save(path)
-        output=lightUp(path)
-        os.remove(path)
+        output, resized = lightUp(path)
+        if resized == None:
+        	flash('The image is too small :( - The minimum height and width are 128 pixels.')
+        	return render_template('lightup-input.html')
         output = Image.fromarray(output.astype("uint8"))
         w, h = output.size
+        if resized:
+        	flash('Warning: Your image was too large and it was resized to width ' + str(w) + ' and height ' +str(h)+'.')
+        os.remove(path)
         width="80%"
         if (h/w>0.5): # Adjust display dimensions
         	width=str(40*w/h)+"%" # Same proportions. Height limited to 40% and width limited to 80%
@@ -68,5 +73,5 @@ def light_up():
 def page_not_found(e):
     return render_template('404.html'), 404
     
-if __name__ == '__main__':
-	app.run(host=HOST, port=PORT_NUMBER, debug=True)
+#if __name__ == '__main__':
+#	app.run(host=HOST, port=PORT_NUMBER, debug=True)
